refactor: extract shared PATH utilities and mobile keyboard hook
