+++ packages/web/index.html
@@ -2,6 +2,8 @@
 <html lang="en" class="h-full">
   <head>
     <meta charset="UTF-8" />
+    <!-- interactive-widget=resizes-content: tells supporting browsers (Chrome 108+, Safari 16.4+) that the
+         page handles keyboard-driven viewport resizing itself. Unsupported browsers ignore the token. -->
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover, interactive-widget=resizes-content" />
 
     <!-- Favicon -->
